Internal changes to prevent `_e4c_library_finalize` from calling `exit` (Fixes  issue 10).
The purpose of the `exit` call was to warn the programmer about a fatal library misuse, by forcing the return value of the program to `EXIT_FAILURE`. Calling `exit` twice may result in undefined behavior; `_e4c_library_finalize` is registered with `atexit`, so we shouldn't be calling `exit` at that point.

According to the standard:

    A function that is registered as an exit handler by `atexit()` must exit by returning, and not in any other manner.

In addition:

    POSIX.1-2001 says that the result of calling exit(3) more than once
    (i.e., calling exit(3) within a function registered using atexit()) is
    undefined. On some systems (but not Linux), this can result in an infinite
    recursion; portable programs should not invoke exit(3) inside a function
    registered using atexit().

The library will not be able to force failure exit status of the program. The existing unit tests must be modified in order to reflect this change.